+++ modules/text/src/test/java/hosh/modules/text/TextModuleTest.java
@@ -22,24 +22,23 @@ package hosh.modules.text;/*
  * SOFTWARE.
  */
 
-import hosh.modules.TextModule.Count;
-import hosh.modules.TextModule.Distinct;
-import hosh.modules.TextModule.Drop;
-import hosh.modules.TextModule.Duplicated;
-import hosh.modules.TextModule.Enumerate;
-import hosh.modules.TextModule.Filter;
-import hosh.modules.TextModule.Join;
-import hosh.modules.TextModule.Regex;
-import hosh.modules.TextModule.Schema;
-import hosh.modules.TextModule.Select;
-import hosh.modules.TextModule.Sort;
-import hosh.modules.TextModule.Split;
-import hosh.modules.TextModule.Sum;
-import hosh.modules.TextModule.Table;
-import hosh.modules.TextModule.Take;
-import hosh.modules.TextModule.Timestamp;
-import hosh.modules.TextModule.Trim;
-import hosh.modules.text.TextModule;
+import hosh.modules.text.TextModule.Count;
+import hosh.modules.text.TextModule.Distinct;
+import hosh.modules.text.TextModule.Drop;
+import hosh.modules.text.TextModule.Duplicated;
+import hosh.modules.text.TextModule.Enumerate;
+import hosh.modules.text.TextModule.Filter;
+import hosh.modules.text.TextModule.Join;
+import hosh.modules.text.TextModule.Regex;
+import hosh.modules.text.TextModule.Schema;
+import hosh.modules.text.TextModule.Select;
+import hosh.modules.text.TextModule.Sort;
+import hosh.modules.text.TextModule.Split;
+import hosh.modules.text.TextModule.Sum;
+import hosh.modules.text.TextModule.Table;
+import hosh.modules.text.TextModule.Take;
+import hosh.modules.text.TextModule.Timestamp;
+import hosh.modules.text.TextModule.Trim;
 import hosh.spi.Ansi;
 import hosh.spi.ExitStatus;
 import hosh.spi.InputChannel;
@@ -48,8 +47,8 @@ import hosh.spi.OutputChannel;
 import hosh.spi.Record;
 import hosh.spi.Records;
 import hosh.spi.Values;
-import hosh.testsupport.RecordMatcher;
-import hosh.testsupport.WithThread;
+import hosh.test.support.RecordMatcher;
+import hosh.test.support.WithThread;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
@@ -68,6 +67,7 @@ import java.time.Instant;
 import java.util.List;
 import java.util.Optional;
 
+import static hosh.test.support.ExitStatusAssert.assertThat;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.mockito.BDDMockito.given;
 import static org.mockito.BDDMockito.then;
@@ -98,7 +98,7 @@ public class TextModuleTest {
 		public void interrupt() {
 			withThread.interrupt();
 			ExitStatus exitStatus = sut.run(List.of(), in, out, err);
-			Assertions.assertThat(exitStatus).isSuccess();
+			assertThat(exitStatus).isSuccess();
 			BDDMockito.then(in).shouldHaveNoMoreInteractions();
 			BDDMockito.then(out).shouldHaveNoInteractions();
 			BDDMockito.then(err).shouldHaveNoInteractions();
@@ -107,7 +107,7 @@ public class TextModuleTest {
 		@Test
 		public void oneArg() {
 			ExitStatus exitStatus = sut.run(List.of("asd"), in, out, err);
-			Assertions.assertThat(exitStatus).isError();
+			assertThat(exitStatus).isError();
 			BDDMockito.then(in).shouldHaveNoMoreInteractions();
 			BDDMockito.then(out).shouldHaveNoInteractions();
 			BDDMockito.then(err).should().send(Records.singleton(Keys.ERROR, Values.ofText("expected 0 arguments")));
@@ -135,7 +135,7 @@ public class TextModuleTest {
 		public void empty() {
 			BDDMockito.given(in.recv()).willReturn(Optional.empty());
 			ExitStatus exitStatus = sut.run(List.of("text"), in, out, err);
-			Assertions.assertThat(exitStatus).isSuccess();
+			assertThat(exitStatus).isSuccess();
 			BDDMockito.then(in).shouldHaveNoMoreInteractions();
 			BDDMockito.then(out).shouldHaveNoInteractions();
 			BDDMockito.then(err).shouldHaveNoInteractions();
@@ -144,7 +144,7 @@ public class TextModuleTest {
 		@Test
 		public void noArgs() {
 			ExitStatus exitStatus = sut.run(List.of(), in, out, err);
-			Assertions.assertThat(exitStatus).isError();
+			assertThat(exitStatus).isError();
 			BDDMockito.then(in).shouldHaveNoMoreInteractions();
 			BDDMockito.then(out).shouldHaveNoInteractions();
 			BDDMockito.then(err).should().send(Records.singleton(Keys.ERROR, Values.ofText("expected 1 argument")));
@@ -156,7 +156,7 @@ public class TextModuleTest {
 			Record record = Records.singleton(Keys.TEXT, Values.ofText("  abc  "));
 			BDDMockito.given(in.recv()).willReturn(Optional.of(record), Optional.empty());
 			ExitStatus exitStatus = sut.run(List.of(Keys.ERROR.name()), in, out, err);
-			Assertions.assertThat(exitStatus).isSuccess();
+			assertThat(exitStatus).isSuccess();
 			BDDMockito.then(in).shouldHaveNoMoreInteractions();
 			BDDMockito.then(out).should().send(record);
 			BDDMockito.then(err).shouldHaveNoInteractions();
@@ -168,7 +168,7 @@ public class TextModuleTest {
 			Record record = Records.singleton(Keys.TEXT, Values.ofText("  abc  "));
 			BDDMockito.given(in.recv()).willReturn(Optional.of(record), Optional.empty());
 			ExitStatus exitStatus = sut.run(List.of(Keys.TEXT.name()), in, out, err);
-			Assertions.assertThat(exitStatus).isSuccess();
+			assertThat(exitStatus).isSuccess();
 			BDDMockito.then(in).shouldHaveNoMoreInteractions();
 			BDDMockito.then(out).should().send(Records.singleton(Keys.TEXT, Values.ofText("abc")));
 			BDDMockito.then(err).shouldHaveNoInteractions();
@@ -180,7 +180,7 @@ public class TextModuleTest {
 			Record record = Records.singleton(Keys.COUNT, Values.ofInstant(Instant.EPOCH));
 			BDDMockito.given(in.recv()).willReturn(Optional.of(record), Optional.empty());
 			ExitStatus exitStatus = sut.run(List.of(Keys.COUNT.name()), in, out, err);
-			Assertions.assertThat(exitStatus).isSuccess();
+			assertThat(exitStatus).isSuccess();
 			BDDMockito.then(in).shouldHaveNoMoreInteractions();
 			BDDMockito.then(out).should().send(record);
 			BDDMockito.then(err).shouldHaveNoInteractions();
@@ -206,7 +206,7 @@ public class TextModuleTest {
 		@Test
 		public void noArgs() {
 			ExitStatus exitStatus = sut.run(List.of(), in, out, err);
-			Assertions.assertThat(exitStatus).isError();
+			assertThat(exitStatus).isError();
 			BDDMockito.then(in).shouldHaveNoMoreInteractions();
 			BDDMockito.then(out).shouldHaveNoInteractions();
 			BDDMockito.then(err).should().send(Records.singleton(Keys.ERROR, Values.ofText("usage: join separator")));
@@ -216,7 +216,7 @@ public class TextModuleTest {
 		public void empty() {
 			BDDMockito.given(in.recv()).willReturn(Optional.empty());
 			ExitStatus exitStatus = sut.run(List.of(","), in, out, err);
-			Assertions.assertThat(exitStatus).isSuccess();
+			assertThat(exitStatus).isSuccess();
 			BDDMockito.then(in).shouldHaveNoMoreInteractions();
 			BDDMockito.then(out).shouldHaveNoInteractions();
 			BDDMockito.then(err).shouldHaveNoInteractions();
@@ -228,7 +228,7 @@ public class TextModuleTest {
 			Record record = Records.singleton(Keys.INDEX, Values.ofNumeric(1));
 			BDDMockito.given(in.recv()).willReturn(Optional.of(record), Optional.empty());
 			ExitStatus exitStatus = sut.run(List.of(","), in, out, err);
-			Assertions.assertThat(exitStatus).isSuccess();
+			assertThat(exitStatus).isSuccess();
 			BDDMockito.then(in).shouldHaveNoMoreInteractions();
 			BDDMockito.then(out).should().send(Records.singleton(Keys.TEXT, Values.ofText("1")));
 			BDDMockito.then(err).shouldHaveNoInteractions();
@@ -240,7 +240,7 @@ public class TextModuleTest {
 			Record record = Records.builder().entry(Keys.INDEX, Values.ofNumeric(1)).entry(Keys.NAME, Values.ofNumeric(2)).build();
 			BDDMockito.given(in.recv()).willReturn(Optional.of(record), Optional.empty());
 			ExitStatus exitStatus = sut.run(List.of(","), in, out, err);
-			Assertions.assertThat(exitStatus).isSuccess();
+			assertThat(exitStatus).isSuccess();
 			BDDMockito.then(in).shouldHaveNoMoreInteractions();
 			BDDMockito.then(out).should().send(Records.singleton(Keys.TEXT, Values.ofText("1,2")));
 			BDDMockito.then(err).shouldHaveNoInteractions();
@@ -266,7 +266,7 @@ public class TextModuleTest {
 		@Test
 		public void noArgs() {
 			ExitStatus exitStatus = sut.run(List.of(), in, out, err);
-			Assertions.assertThat(exitStatus).isError();
+			assertThat(exitStatus).isError();
 			BDDMockito.then(in).shouldHaveNoMoreInteractions();
 			BDDMockito.then(out).shouldHaveNoInteractions();
 			BDDMockito.then(err).should().send(Records.singleton(Keys.ERROR, Values.ofText("usage: sum key")));
@@ -276,7 +276,7 @@ public class TextModuleTest {
 		public void empty() {
 			BDDMockito.given(in.recv()).willReturn(Optional.empty());
 			ExitStatus exitStatus = sut.run(List.of("size"), in, out, err);
-			Assertions.assertThat(exitStatus).isSuccess();
+			assertThat(exitStatus).isSuccess();
 			BDDMockito.then(in).shouldHaveNoMoreInteractions();
 			BDDMockito.then(out).should().send(RecordMatcher.of(Keys.SIZE, Values.ofSize(0)));
 			BDDMockito.then(err).shouldHaveNoInteractions();
@@ -288,7 +288,7 @@ public class TextModuleTest {
 			Record record = Records.singleton(Keys.INDEX, Values.ofSize(1));
 			BDDMockito.given(in.recv()).willReturn(Optional.of(record), Optional.empty());
 			ExitStatus exitStatus = sut.run(List.of("size"), in, out, err);
-			Assertions.assertThat(exitStatus).isSuccess();
+			assertThat(exitStatus).isSuccess();
 			BDDMockito.then(in).shouldHaveNoMoreInteractions();
 			BDDMockito.then(out).should().send(RecordMatcher.of(Keys.SIZE, Values.ofSize(0)));
 			BDDMockito.then(err).shouldHaveNoInteractions();
@@ -300,7 +300,7 @@ public class TextModuleTest {
 			Record record = Records.singleton(Keys.SIZE, Values.ofSize(1));
 			BDDMockito.given(in.recv()).willReturn(Optional.of(record), Optional.of(record), Optional.empty());
 			ExitStatus exitStatus = sut.run(List.of("size"), in, out, err);
-			Assertions.assertThat(exitStatus).isSuccess();
+			assertThat(exitStatus).isSuccess();
 			BDDMockito.then(in).shouldHaveNoMoreInteractions();
 			BDDMockito.then(out).should().send(RecordMatcher.of(Keys.SIZE, Values.ofSize(2)));
 			BDDMockito.then(err).shouldHaveNoInteractions();
@@ -327,7 +327,7 @@ public class TextModuleTest {
 		public void empty() {
 			BDDMockito.given(in.recv()).willReturn(Optional.empty());
 			ExitStatus exitStatus = sut.run(List.of(Keys.COUNT.name()), in, out, err);
-			Assertions.assertThat(exitStatus).isSuccess();
+			assertThat(exitStatus).isSuccess();
 			BDDMockito.then(in).shouldHaveNoMoreInteractions();
 			BDDMockito.then(out).shouldHaveNoInteractions();
 			BDDMockito.then(err).shouldHaveNoInteractions();
@@ -338,7 +338,7 @@ public class TextModuleTest {
 		public void noArgs() {
 			BDDMockito.given(in.recv()).willReturn(Optional.of(Records.singleton(Keys.NAME, Values.ofNumeric(1))), Optional.empty());
 			ExitStatus exitStatus = sut.run(List.of(), in, out, err);
-			Assertions.assertThat(exitStatus).isSuccess();
+			assertThat(exitStatus).isSuccess();
 			BDDMockito.then(in).shouldHaveNoMoreInteractions();
 			BDDMockito.then(out).should().send(Records.empty());
 			BDDMockito.then(err).shouldHaveNoInteractions();
@@ -350,7 +350,7 @@ public class TextModuleTest {
 			Record record = Records.singleton(Keys.NAME, Values.ofText("foo"));
 			BDDMockito.given(in.recv()).willReturn(Optional.of(record), Optional.empty());
 			ExitStatus exitStatus = sut.run(List.of(Keys.NAME.name()), in, out, err);
-			Assertions.assertThat(exitStatus).isSuccess();
+			assertThat(exitStatus).isSuccess();
 			BDDMockito.then(in).should(times(2)).recv();
 			BDDMockito.then(out).should().send(record);
 			BDDMockito.then(err).shouldHaveNoInteractions();
@@ -362,7 +362,7 @@ public class TextModuleTest {
 			Record record = Records.singleton(Keys.NAME, Values.ofText("foo"));
 			BDDMockito.given(in.recv()).willReturn(Optional.of(record), Optional.empty());
 			ExitStatus exitStatus = sut.run(List.of(Keys.NAME.name(), Keys.COUNT.name()), in, out, err);
-			Assertions.assertThat(exitStatus).isSuccess();
+			assertThat(exitStatus).isSuccess();
 			BDDMockito.then(in).should(times(2)).recv();
 			BDDMockito.then(out).should().send(record);
 			BDDMockito.then(err).shouldHaveNoInteractions();
@@ -388,7 +388,7 @@ public class TextModuleTest {
 		@Test
 		public void noArgs() {
 			ExitStatus exitStatus = sut.run(List.of(), in, out, err);
-			Assertions.assertThat(exitStatus).isError();
+			assertThat(exitStatus).isError();
 			BDDMockito.then(in).shouldHaveNoInteractions();
 			BDDMockito.then(out).shouldHaveNoInteractions();
 			BDDMockito.then(err).should().send(Records.singleton(Keys.ERROR, Values.ofText("usage: split key regex")));
@@ -397,7 +397,7 @@ public class TextModuleTest {
 		@Test
 		public void oneArg() {
 			ExitStatus exitStatus = sut.run(List.of(Keys.TEXT.name()), in, out, err);
-			Assertions.assertThat(exitStatus).isError();
+			assertThat(exitStatus).isError();
 			BDDMockito.then(in).shouldHaveNoInteractions();
 			BDDMockito.then(out).shouldHaveNoInteractions();
 			BDDMockito.then(err).should().send(Records.singleton(Keys.ERROR, Values.ofText("usage: split key regex")));
@@ -407,7 +407,7 @@ public class TextModuleTest {
 		public void twoArgsNoInput() {
 			BDDMockito.given(in.recv()).willReturn(Optional.empty());
 			ExitStatus exitStatus = sut.run(List.of(Keys.TEXT.name(), " "), in, out, err);
-			Assertions.assertThat(exitStatus).isSuccess();
+			assertThat(exitStatus).isSuccess();
 			BDDMockito.then(in).shouldHaveNoMoreInteractions();
 			BDDMockito.then(out).shouldHaveNoInteractions();
 			BDDMockito.then(err).shouldHaveNoInteractions();
@@ -419,7 +419,7 @@ public class TextModuleTest {
 			Record record = Records.singleton(Keys.TEXT, Values.ofText("a b c"));
 			BDDMockito.given(in.recv()).willReturn(Optional.of(record), Optional.empty());
 			ExitStatus exitStatus = sut.run(List.of(Keys.TEXT.name(), " "), in, out, err);
-			Assertions.assertThat(exitStatus).isSuccess();
+			assertThat(exitStatus).isSuccess();
 			BDDMockito.then(in).should(times(2)).recv();
 			BDDMockito.then(out).should().send(Records.builder()
 					.entry(Keys.of("1"), Values.ofText("a"))
@@ -449,7 +449,7 @@ public class TextModuleTest {
 		@Test
 		public void zeroArg() {
 			ExitStatus exitStatus = sut.run(List.of(), in, out, err);
-			Assertions.assertThat(exitStatus).isError();
+			assertThat(exitStatus).isError();
 			BDDMockito.then(in).shouldHaveNoInteractions();
 			BDDMockito.then(out).shouldHaveNoInteractions();
 			BDDMockito.then(err).should().send(Records.singleton(Keys.ERROR, Values.ofText("expected 2 arguments")));
@@ -458,7 +458,7 @@ public class TextModuleTest {
 		@Test
 		public void oneArg() {
 			ExitStatus exitStatus = sut.run(List.of("asd"), in, out, err);
-			Assertions.assertThat(exitStatus).isError();
+			assertThat(exitStatus).isError();
 			BDDMockito.then(in).shouldHaveNoInteractions();
 			BDDMockito.then(out).shouldHaveNoInteractions();
 			BDDMockito.then(err).should().send(Records.singleton(Keys.ERROR, Values.ofText("expected 2 arguments")));
@@ -468,7 +468,7 @@ public class TextModuleTest {
 		public void twoArgsNoInput() {
 			BDDMockito.given(in.recv()).willReturn(Optional.empty());
 			ExitStatus exitStatus = sut.run(List.of(Keys.TEXT.name(), "(?<id>\\\\d+)"), in, out, err);
-			Assertions.assertThat(exitStatus).isSuccess();
+			assertThat(exitStatus).isSuccess();
 			BDDMockito.then(in).shouldHaveNoMoreInteractions();
 			BDDMockito.then(out).shouldHaveNoInteractions();
 			BDDMockito.then(err).shouldHaveNoInteractions();
@@ -480,7 +480,7 @@ public class TextModuleTest {
 			Record record = Records.singleton(Keys.TEXT, Values.ofText(""));
 			BDDMockito.given(in.recv()).willReturn(Optional.of(record), Optional.empty());
 			ExitStatus exitStatus = sut.run(List.of(Keys.TEXT.name(), "(?<id>\\\\d+)"), in, out, err);
-			Assertions.assertThat(exitStatus).isSuccess();
+			assertThat(exitStatus).isSuccess();
 			BDDMockito.then(in).shouldHaveNoMoreInteractions();
 			BDDMockito.then(out).shouldHaveNoInteractions();
 			BDDMockito.then(err).shouldHaveNoInteractions();
@@ -492,7 +492,7 @@ public class TextModuleTest {
 			Record record = Records.singleton(Keys.TEXT, Values.ofText("1"));
 			BDDMockito.given(in.recv()).willReturn(Optional.of(record), Optional.empty());
 			ExitStatus exitStatus = sut.run(List.of(Keys.TEXT.name(), "(?<id>\\d+)"), in, out, err);
-			Assertions.assertThat(exitStatus).isSuccess();
+			assertThat(exitStatus).isSuccess();
 			BDDMockito.then(in).should(times(2)).recv();
 			BDDMockito.then(out).should().send(Records.singleton(Keys.of("id"), Values.ofText("1")));
 			BDDMockito.then(err).shouldHaveNoInteractions();
@@ -504,7 +504,7 @@ public class TextModuleTest {
 			Record record = Records.singleton(Keys.TEXT, Values.ofText("1"));
 			BDDMockito.given(in.recv()).willReturn(Optional.of(record), Optional.empty());
 			ExitStatus exitStatus = sut.run(List.of(Keys.COUNT.name(), "(?<id>\\d+)"), in, out, err);
-			Assertions.assertThat(exitStatus).isSuccess();
+			assertThat(exitStatus).isSuccess();
 			BDDMockito.then(in).should(times(2)).recv();
 			BDDMockito.then(out).shouldHaveNoInteractions();
 			BDDMockito.then(err).shouldHaveNoInteractions();
@@ -533,7 +533,7 @@ public class TextModuleTest {
 			Record record = Records.singleton(Keys.COUNT, null).append(Keys.INDEX, null);
 			BDDMockito.given(in.recv()).willReturn(Optional.of(record), Optional.empty());
 			ExitStatus exitStatus = sut.run(List.of(), in, out, err);
-			Assertions.assertThat(exitStatus).isSuccess();
+			assertThat(exitStatus).isSuccess();
 			BDDMockito.then(in).shouldHaveNoMoreInteractions();
 			BDDMockito.then(out).should().send(Records.singleton(Keys.of("schema"), Values.ofText("count index")));
 			BDDMockito.then(err).shouldHaveNoMoreInteractions();
@@ -542,7 +542,7 @@ public class TextModuleTest {
 		@Test
 		public void oneArg() {
 			ExitStatus exitStatus = sut.run(List.of("asd"), in, out, err);
-			Assertions.assertThat(exitStatus).isError();
+			assertThat(exitStatus).isError();
 			BDDMockito.then(in).shouldHaveNoInteractions();
 			BDDMockito.then(out).shouldHaveNoMoreInteractions();
 			BDDMockito.then(err).should().send(Records.singleton(Keys.ERROR, Values.ofText("expected 0 arguments")));
@@ -572,7 +572,7 @@ public class TextModuleTest {
 			Record record = Records.singleton(Keys.TEXT, Values.ofText("some data"));
 			BDDMockito.given(in.recv()).willReturn(Optional.of(record), Optional.of(record), Optional.empty());
 			ExitStatus exitStatus = sut.run(List.of(), in, out, err);
-			Assertions.assertThat(exitStatus).isSuccess();
+			assertThat(exitStatus).isSuccess();
 			BDDMockito.then(in).shouldHaveNoMoreInteractions();
 			BDDMockito.then(out).should().send(Records.singleton(Keys.COUNT, Values.ofNumeric(2)));
 			BDDMockito.then(err).shouldHaveNoMoreInteractions();
@@ -584,7 +584,7 @@ public class TextModuleTest {
 			Record record = Records.singleton(Keys.TEXT, Values.ofText("some data"));
 			BDDMockito.given(in.recv()).willReturn(Optional.of(record), Optional.empty());
 			ExitStatus exitStatus = sut.run(List.of(), in, out, err);
-			Assertions.assertThat(exitStatus).isSuccess();
+			assertThat(exitStatus).isSuccess();
 			BDDMockito.then(in).shouldHaveNoMoreInteractions();
 			BDDMockito.then(out).should().send(Records.singleton(Keys.COUNT, Values.ofNumeric(1)));
 			BDDMockito.then(err).shouldHaveNoMoreInteractions();
@@ -594,7 +594,7 @@ public class TextModuleTest {
 		public void zeroRecords() {
 			BDDMockito.given(in.recv()).willReturn(Optional.empty());
 			ExitStatus exitStatus = sut.run(List.of(), in, out, err);
-			Assertions.assertThat(exitStatus).isSuccess();
+			assertThat(exitStatus).isSuccess();
 			BDDMockito.then(in).shouldHaveNoMoreInteractions();
 			BDDMockito.then(out).should().send(Records.singleton(Keys.COUNT, Values.ofNumeric(0)));
 			BDDMockito.then(err).shouldHaveNoMoreInteractions();
@@ -603,7 +603,7 @@ public class TextModuleTest {
 		@Test
 		public void oneArg() {
 			ExitStatus exitStatus = sut.run(List.of("asd"), in, out, err);
-			Assertions.assertThat(exitStatus).isError();
+			assertThat(exitStatus).isError();
 			BDDMockito.then(in).shouldHaveNoInteractions();
 			BDDMockito.then(out).shouldHaveNoMoreInteractions();
 			BDDMockito.then(err).should().send(Records.singleton(Keys.ERROR, Values.ofText("expected 0 arguments")));
@@ -633,7 +633,7 @@ public class TextModuleTest {
 			Record record = Records.singleton(Keys.TEXT, Values.ofText("some data"));
 			BDDMockito.given(in.recv()).willReturn(Optional.of(record), Optional.of(Records.empty()), Optional.empty());
 			ExitStatus exitStatus = sut.run(List.of(), in, out, err);
-			Assertions.assertThat(exitStatus).isSuccess();
+			assertThat(exitStatus).isSuccess();
 			BDDMockito.then(in).shouldHaveNoMoreInteractions();
 			BDDMockito.then(out).should().send(Records.builder().entry(Keys.INDEX, Values.ofNumeric(1)).entry(Keys.TEXT, Values.ofText("some data")).build());
 			BDDMockito.then(out).should().send(Records.singleton(Keys.INDEX, Values.ofNumeric(2)));
@@ -643,7 +643,7 @@ public class TextModuleTest {
 		@Test
 		public void oneArg() {
 			ExitStatus exitStatus = sut.run(List.of("asd"), in, out, err);
-			Assertions.assertThat(exitStatus).isError();
+			assertThat(exitStatus).isError();
 			BDDMockito.then(in).shouldHaveNoInteractions();
 			BDDMockito.then(out).shouldHaveNoMoreInteractions();
 			BDDMockito.then(err).should().send(Records.singleton(Keys.ERROR, Values.ofText("expected 0 arguments")));
@@ -677,7 +677,7 @@ public class TextModuleTest {
 			Record record = Records.singleton(Keys.TEXT, Values.ofText("some data"));
 			BDDMockito.given(in.recv()).willReturn(Optional.of(record), Optional.of(Records.empty()), Optional.empty());
 			ExitStatus exitStatus = sut.run(List.of(), in, out, err);
-			Assertions.assertThat(exitStatus).isSuccess();
+			assertThat(exitStatus).isSuccess();
 			BDDMockito.then(in).shouldHaveNoMoreInteractions();
 			BDDMockito.then(out).should().send(
 					Records.builder()
@@ -690,7 +690,7 @@ public class TextModuleTest {
 		@Test
 		public void oneArg() {
 			ExitStatus exitStatus = sut.run(List.of("asd"), in, out, err);
-			Assertions.assertThat(exitStatus).isError();
+			assertThat(exitStatus).isError();
 			BDDMockito.then(in).shouldHaveNoInteractions();
 			BDDMockito.then(out).shouldHaveNoMoreInteractions();
 			BDDMockito.then(err).should().send(Records.singleton(Keys.ERROR, Values.ofText("expected 0 arguments")));
@@ -720,7 +720,7 @@ public class TextModuleTest {
 			Record record = Records.singleton(Keys.TEXT, Values.ofText("some data"));
 			BDDMockito.given(in.recv()).willReturn(Optional.of(record), Optional.empty());
 			ExitStatus exitStatus = sut.run(List.of("0"), in, out, err);
-			Assertions.assertThat(exitStatus).isSuccess();
+			assertThat(exitStatus).isSuccess();
 			BDDMockito.then(in).shouldHaveNoMoreInteractions();
 			BDDMockito.then(out).should().send(record);
 			BDDMockito.then(err).shouldHaveNoMoreInteractions();
@@ -732,7 +732,7 @@ public class TextModuleTest {
 			Record record = Records.singleton(Keys.TEXT, Values.ofText("some data"));
 			BDDMockito.given(in.recv()).willReturn(Optional.of(record), Optional.empty());
 			ExitStatus exitStatus = sut.run(List.of("1"), in, out, err);
-			Assertions.assertThat(exitStatus).isSuccess();
+			assertThat(exitStatus).isSuccess();
 			BDDMockito.then(in).shouldHaveNoMoreInteractions();
 			BDDMockito.then(out).shouldHaveNoMoreInteractions();
 			BDDMockito.then(err).shouldHaveNoMoreInteractions();
@@ -741,7 +741,7 @@ public class TextModuleTest {
 		@Test
 		public void noArgs() {
 			ExitStatus exitStatus = sut.run(List.of(), in, out, err);
-			Assertions.assertThat(exitStatus).isError();
+			assertThat(exitStatus).isError();
 			BDDMockito.then(in).shouldHaveNoInteractions();
 			BDDMockito.then(out).shouldHaveNoMoreInteractions();
 			BDDMockito.then(err).should().send(Records.singleton(Keys.ERROR, Values.ofText("expected 1 parameter")));
@@ -751,7 +751,7 @@ public class TextModuleTest {
 		@Test
 		public void negativeArg() {
 			ExitStatus exitStatus = sut.run(List.of("-1"), in, out, err);
-			Assertions.assertThat(exitStatus).isError();
+			assertThat(exitStatus).isError();
 			BDDMockito.then(in).shouldHaveNoInteractions();
 			BDDMockito.then(out).shouldHaveNoInteractions();
 			BDDMockito.then(err).should().send(Records.singleton(Keys.ERROR, Values.ofText("parameter must be >= 0")));
@@ -778,7 +778,7 @@ public class TextModuleTest {
 		public void takeZero() {
 			BDDMockito.given(in.recv()).willReturn(Optional.empty());
 			ExitStatus exitStatus = sut.run(List.of("0"), in, out, err);
-			Assertions.assertThat(exitStatus).isSuccess();
+			assertThat(exitStatus).isSuccess();
 			BDDMockito.then(in).should().recv();
 			BDDMockito.then(out).shouldHaveNoInteractions();
 			BDDMockito.then(err).shouldHaveNoMoreInteractions();
@@ -790,7 +790,7 @@ public class TextModuleTest {
 			Record record = Records.singleton(Keys.TEXT, Values.ofText("some data"));
 			BDDMockito.given(in.recv()).willReturn(Optional.of(record), Optional.empty());
 			ExitStatus exitStatus = sut.run(List.of("1"), in, out, err);
-			Assertions.assertThat(exitStatus).isSuccess();
+			assertThat(exitStatus).isSuccess();
 			BDDMockito.then(in).shouldHaveNoMoreInteractions();
 			BDDMockito.then(out).should().send(record);
 			BDDMockito.then(err).shouldHaveNoMoreInteractions();
@@ -802,7 +802,7 @@ public class TextModuleTest {
 			Record record = Records.singleton(Keys.TEXT, Values.ofText("some data"));
 			BDDMockito.given(in.recv()).willReturn(Optional.of(record), Optional.of(record), Optional.empty());
 			ExitStatus exitStatus = sut.run(List.of("1"), in, out, err);
-			Assertions.assertThat(exitStatus).isSuccess();
+			assertThat(exitStatus).isSuccess();
 			BDDMockito.then(in).shouldHaveNoMoreInteractions();
 			BDDMockito.then(out).should().send(record);
 			BDDMockito.then(err).shouldHaveNoMoreInteractions();
@@ -815,7 +815,7 @@ public class TextModuleTest {
 			Record record2 = Records.singleton(Keys.TEXT, Values.ofText("another value"));
 			BDDMockito.given(in.recv()).willReturn(Optional.of(record), Optional.of(record2), Optional.empty());
 			ExitStatus exitStatus = sut.run(List.of("5"), in, out, err);
-			Assertions.assertThat(exitStatus).isSuccess();
+			assertThat(exitStatus).isSuccess();
 			BDDMockito.then(in).shouldHaveNoMoreInteractions();
 			BDDMockito.then(out).should().send(record);
 			BDDMockito.then(out).should().send(record2);
@@ -826,7 +826,7 @@ public class TextModuleTest {
 		@Test
 		public void negativeArg() {
 			ExitStatus exitStatus = sut.run(List.of("-1"), in, out, err);
-			Assertions.assertThat(exitStatus).isError();
+			assertThat(exitStatus).isError();
 			BDDMockito.then(in).shouldHaveNoInteractions();
 			BDDMockito.then(out).shouldHaveNoInteractions();
 			BDDMockito.then(err).should().send(Records.singleton(Keys.ERROR, Values.ofText("parameter must be >= 0")));
@@ -835,7 +835,7 @@ public class TextModuleTest {
 		@Test
 		public void noArgs() {
 			ExitStatus exitStatus = sut.run(List.of(), in, out, err);
-			Assertions.assertThat(exitStatus).isError();
+			assertThat(exitStatus).isError();
 			BDDMockito.then(out).shouldHaveNoInteractions();
 			BDDMockito.then(out).shouldHaveNoInteractions();
 			BDDMockito.then(err).should().send(Records.singleton(Keys.ERROR, Values.ofText("expected 1 parameter")));
@@ -865,7 +865,7 @@ public class TextModuleTest {
 			Record record = Records.singleton(Keys.TEXT, Values.ofText("some string"));
 			BDDMockito.given(in.recv()).willReturn(Optional.of(record), Optional.empty());
 			ExitStatus exitStatus = sut.run(List.of(Keys.TEXT.name(), ".*string.*"), in, out, err);
-			Assertions.assertThat(exitStatus).isSuccess();
+			assertThat(exitStatus).isSuccess();
 			BDDMockito.then(in).shouldHaveNoMoreInteractions();
 			BDDMockito.then(out).should().send(record);
 			BDDMockito.then(err).shouldHaveNoMoreInteractions();
@@ -877,7 +877,7 @@ public class TextModuleTest {
 			Record record = Records.singleton(Keys.TEXT, Values.ofText("some string"));
 			BDDMockito.given(in.recv()).willReturn(Optional.of(record), Optional.empty());
 			ExitStatus exitStatus = sut.run(List.of("key", ".*number.*"), in, out, err);
-			Assertions.assertThat(exitStatus).isSuccess();
+			assertThat(exitStatus).isSuccess();
 			BDDMockito.then(in).shouldHaveNoMoreInteractions();
 			BDDMockito.then(out).shouldHaveNoInteractions();
 			BDDMockito.then(err).shouldHaveNoMoreInteractions();
@@ -886,7 +886,7 @@ public class TextModuleTest {
 		@Test
 		public void zeroArgs() {
 			ExitStatus exitStatus = sut.run(List.of(), in, out, err);
-			Assertions.assertThat(exitStatus).isError();
+			assertThat(exitStatus).isError();
 			BDDMockito.then(in).shouldHaveNoInteractions();
 			BDDMockito.then(out).shouldHaveNoInteractions();
 			BDDMockito.then(err).should().send(Records.singleton(Keys.ERROR, Values.ofText("expected 2 arguments: key regex")));
@@ -896,7 +896,7 @@ public class TextModuleTest {
 		@Test
 		public void oneArg() {
 			ExitStatus exitStatus = sut.run(List.of("key"), in, out, err);
-			Assertions.assertThat(exitStatus).isError();
+			assertThat(exitStatus).isError();
 			BDDMockito.then(in).shouldHaveNoInteractions();
 			BDDMockito.then(out).shouldHaveNoMoreInteractions();
 			BDDMockito.then(err).should().send(Records.singleton(Keys.ERROR, Values.ofText("expected 2 arguments: key regex")));
@@ -927,7 +927,7 @@ public class TextModuleTest {
 		public void empty() {
 			BDDMockito.given(in.recv()).willReturn(Optional.empty());
 			ExitStatus exitStatus = sut.run(List.of("name"), in, out, err);
-			Assertions.assertThat(exitStatus).isSuccess();
+			assertThat(exitStatus).isSuccess();
 			BDDMockito.then(in).shouldHaveNoMoreInteractions();
 			BDDMockito.then(err).shouldHaveNoMoreInteractions();
 			BDDMockito.then(out).shouldHaveNoInteractions();
@@ -940,11 +940,11 @@ public class TextModuleTest {
 			Record record2 = Records.singleton(Keys.NAME, Values.ofText("aaa"));
 			BDDMockito.given(in.recv()).willReturn(Optional.of(record1), Optional.of(record2), Optional.empty());
 			ExitStatus exitStatus = sut.run(List.of("name"), in, out, err);
-			Assertions.assertThat(exitStatus).isSuccess();
+			assertThat(exitStatus).isSuccess();
 			BDDMockito.then(in).shouldHaveNoMoreInteractions();
 			BDDMockito.then(err).shouldHaveNoMoreInteractions();
 			BDDMockito.then(out).should(times(2)).send(records.capture());
-			Assertions.assertThat(records.getAllValues()).containsExactly(record2, record1);
+			assertThat(records.getAllValues()).containsExactly(record2, record1);
 		}
 
 		@SuppressWarnings("unchecked")
@@ -955,11 +955,11 @@ public class TextModuleTest {
 			Record record3 = Records.singleton(Keys.NAME, Values.ofText("aaa"));
 			BDDMockito.given(in.recv()).willReturn(Optional.of(record1), Optional.of(record2), Optional.of(record3), Optional.empty());
 			ExitStatus exitStatus = sut.run(List.of("name"), in, out, err);
-			Assertions.assertThat(exitStatus).isSuccess();
+			assertThat(exitStatus).isSuccess();
 			BDDMockito.then(in).shouldHaveNoMoreInteractions();
 			BDDMockito.then(err).shouldHaveNoMoreInteractions();
 			BDDMockito.then(out).should(times(3)).send(records.capture());
-			Assertions.assertThat(records.getAllValues()).containsExactly(record3, record2, record1);
+			assertThat(records.getAllValues()).containsExactly(record3, record2, record1);
 		}
 
 		@SuppressWarnings("unchecked")
@@ -968,11 +968,11 @@ public class TextModuleTest {
 			Record record1 = Records.singleton(Keys.NAME, Values.ofText("aaa"));
 			BDDMockito.given(in.recv()).willReturn(Optional.of(record1), Optional.empty());
 			ExitStatus exitStatus = sut.run(List.of("size"), in, out, err);
-			Assertions.assertThat(exitStatus).isSuccess();
+			assertThat(exitStatus).isSuccess();
 			BDDMockito.then(in).shouldHaveNoMoreInteractions();
 			BDDMockito.then(err).shouldHaveNoMoreInteractions();
 			BDDMockito.then(out).should(times(1)).send(records.capture());
-			Assertions.assertThat(records.getAllValues()).containsExactly(record1);
+			assertThat(records.getAllValues()).containsExactly(record1);
 		}
 
 		@SuppressWarnings("unchecked")
@@ -982,11 +982,11 @@ public class TextModuleTest {
 			Record record2 = Records.singleton(Keys.NAME, Values.ofText("aaa"));
 			BDDMockito.given(in.recv()).willReturn(Optional.of(record1), Optional.of(record2), Optional.empty());
 			ExitStatus exitStatus = sut.run(List.of("asc", "name"), in, out, err);
-			Assertions.assertThat(exitStatus).isSuccess();
+			assertThat(exitStatus).isSuccess();
 			BDDMockito.then(in).shouldHaveNoMoreInteractions();
 			BDDMockito.then(err).shouldHaveNoMoreInteractions();
 			BDDMockito.then(out).should(times(2)).send(records.capture());
-			Assertions.assertThat(records.getAllValues()).containsExactly(record2, record1);
+			assertThat(records.getAllValues()).containsExactly(record2, record1);
 		}
 
 		@SuppressWarnings("unchecked")
@@ -996,17 +996,17 @@ public class TextModuleTest {
 			Record record2 = Records.singleton(Keys.NAME, Values.ofText("aaa"));
 			BDDMockito.given(in.recv()).willReturn(Optional.of(record1), Optional.of(record2), Optional.empty());
 			ExitStatus exitStatus = sut.run(List.of("desc", "name"), in, out, err);
-			Assertions.assertThat(exitStatus).isSuccess();
+			assertThat(exitStatus).isSuccess();
 			BDDMockito.then(in).shouldHaveNoMoreInteractions();
 			BDDMockito.then(err).shouldHaveNoMoreInteractions();
 			BDDMockito.then(out).should(times(2)).send(records.capture());
-			Assertions.assertThat(records.getAllValues()).containsExactly(record1, record2);
+			assertThat(records.getAllValues()).containsExactly(record1, record2);
 		}
 
 		@Test
 		public void invalidDirection() {
 			ExitStatus exitStatus = sut.run(List.of("ZZZ", "name"), in, out, err);
-			Assertions.assertThat(exitStatus).isError();
+			assertThat(exitStatus).isError();
 			BDDMockito.then(in).shouldHaveNoMoreInteractions();
 			BDDMockito.then(out).shouldHaveNoMoreInteractions();
 			BDDMockito.then(err).should().send(Records.singleton(Keys.ERROR, Values.ofText("must be asc or desc")));
@@ -1016,7 +1016,7 @@ public class TextModuleTest {
 		@Test
 		public void zeroArgs() {
 			ExitStatus exitStatus = sut.run(List.of(), in, out, err);
-			Assertions.assertThat(exitStatus).isError();
+			assertThat(exitStatus).isError();
 			BDDMockito.then(in).shouldHaveNoMoreInteractions();
 			BDDMockito.then(out).shouldHaveNoMoreInteractions();
 			BDDMockito.then(err).should().send(Records.singleton(Keys.ERROR, Values.ofText("use 'sort key' or 'sort [asc|desc] key'")));
@@ -1026,7 +1026,7 @@ public class TextModuleTest {
 		@Test
 		public void tooManyArgs() {
 			ExitStatus exitStatus = sut.run(List.of("asc", "key", "aaa"), in, out, err);
-			Assertions.assertThat(exitStatus).isError();
+			assertThat(exitStatus).isError();
 			BDDMockito.then(in).shouldHaveNoMoreInteractions();
 			BDDMockito.then(out).shouldHaveNoMoreInteractions();
 			BDDMockito.then(err).should().send(Records.singleton(Keys.ERROR, Values.ofText("too many args")));
@@ -1060,11 +1060,11 @@ public class TextModuleTest {
 			Record record2 = Records.singleton(Keys.NAME, Values.ofNumeric(1));
 			BDDMockito.given(in.recv()).willReturn(Optional.of(record1), Optional.of(record2), Optional.empty());
 			ExitStatus exitStatus = sut.run(List.of("anotherkey"), in, out, err);
-			Assertions.assertThat(exitStatus).isSuccess();
+			assertThat(exitStatus).isSuccess();
 			BDDMockito.then(in).shouldHaveNoMoreInteractions();
 			BDDMockito.then(err).shouldHaveNoMoreInteractions();
 			BDDMockito.then(out).should(Mockito.never()).send(records.capture());
-			Assertions.assertThat(records.getAllValues()).isEmpty();
+			assertThat(records.getAllValues()).isEmpty();
 		}
 
 		@SuppressWarnings("unchecked")
@@ -1074,11 +1074,11 @@ public class TextModuleTest {
 			Record record2 = Records.singleton(Keys.NAME, Values.ofNumeric(1));
 			BDDMockito.given(in.recv()).willReturn(Optional.of(record1), Optional.of(record2), Optional.empty());
 			ExitStatus exitStatus = sut.run(List.of("name"), in, out, err);
-			Assertions.assertThat(exitStatus).isSuccess();
+			assertThat(exitStatus).isSuccess();
 			BDDMockito.then(in).shouldHaveNoMoreInteractions();
 			BDDMockito.then(err).shouldHaveNoMoreInteractions();
 			BDDMockito.then(out).should(times(2)).send(records.capture());
-			Assertions.assertThat(records.getAllValues()).containsExactlyInAnyOrder(record1, record2);
+			assertThat(records.getAllValues()).containsExactlyInAnyOrder(record1, record2);
 		}
 
 		@SuppressWarnings("unchecked")
@@ -1088,17 +1088,17 @@ public class TextModuleTest {
 			Record record2 = Records.singleton(Keys.NAME, Values.ofNumeric(1));
 			BDDMockito.given(in.recv()).willReturn(Optional.of(record1), Optional.of(record2), Optional.empty());
 			ExitStatus exitStatus = sut.run(List.of("name"), in, out, err);
-			Assertions.assertThat(exitStatus).isSuccess();
+			assertThat(exitStatus).isSuccess();
 			BDDMockito.then(in).shouldHaveNoMoreInteractions();
 			BDDMockito.then(err).shouldHaveNoMoreInteractions();
 			BDDMockito.then(out).should().send(records.capture());
-			Assertions.assertThat(records.getAllValues()).containsExactlyInAnyOrder(record1);
+			assertThat(records.getAllValues()).containsExactlyInAnyOrder(record1);
 		}
 
 		@Test
 		public void zeroArgs() {
 			ExitStatus exitStatus = sut.run(List.of(), in, out, err);
-			Assertions.assertThat(exitStatus).isError();
+			assertThat(exitStatus).isError();
 			BDDMockito.then(in).shouldHaveNoMoreInteractions();
 			BDDMockito.then(out).shouldHaveNoMoreInteractions();
 			BDDMockito.then(err).should().send(Records.singleton(Keys.ERROR, Values.ofText("expected 1 parameter: key")));
@@ -1132,11 +1132,11 @@ public class TextModuleTest {
 			Record record2 = Records.singleton(Keys.NAME, Values.ofNumeric(1));
 			BDDMockito.given(in.recv()).willReturn(Optional.of(record1), Optional.of(record2), Optional.empty());
 			ExitStatus exitStatus = sut.run(List.of("anotherkey"), in, out, err);
-			Assertions.assertThat(exitStatus).isSuccess();
+			assertThat(exitStatus).isSuccess();
 			BDDMockito.then(in).shouldHaveNoMoreInteractions();
 			BDDMockito.then(err).shouldHaveNoMoreInteractions();
 			BDDMockito.then(out).should(times(0)).send(records.capture());
-			Assertions.assertThat(records.getAllValues()).isEmpty();
+			assertThat(records.getAllValues()).isEmpty();
 		}
 
 		@SuppressWarnings("unchecked")
@@ -1146,11 +1146,11 @@ public class TextModuleTest {
 			Record record2 = Records.singleton(Keys.NAME, Values.ofNumeric(1));
 			BDDMockito.given(in.recv()).willReturn(Optional.of(record1), Optional.of(record2), Optional.empty());
 			ExitStatus exitStatus = sut.run(List.of("name"), in, out, err);
-			Assertions.assertThat(exitStatus).isSuccess();
+			assertThat(exitStatus).isSuccess();
 			BDDMockito.then(in).shouldHaveNoMoreInteractions();
 			BDDMockito.then(err).shouldHaveNoMoreInteractions();
 			BDDMockito.then(out).should(Mockito.never()).send(records.capture());
-			Assertions.assertThat(records.getAllValues()).isEmpty();
+			assertThat(records.getAllValues()).isEmpty();
 		}
 
 		@SuppressWarnings("unchecked")
@@ -1160,17 +1160,17 @@ public class TextModuleTest {
 			Record record2 = Records.singleton(Keys.NAME, Values.ofNumeric(1));
 			BDDMockito.given(in.recv()).willReturn(Optional.of(record1), Optional.of(record2), Optional.empty());
 			ExitStatus exitStatus = sut.run(List.of("name"), in, out, err);
-			Assertions.assertThat(exitStatus).isSuccess();
+			assertThat(exitStatus).isSuccess();
 			BDDMockito.then(in).shouldHaveNoMoreInteractions();
 			BDDMockito.then(err).shouldHaveNoMoreInteractions();
 			BDDMockito.then(out).should(times(1)).send(records.capture());
-			Assertions.assertThat(records.getAllValues()).containsExactlyInAnyOrder(record1);
+			assertThat(records.getAllValues()).containsExactlyInAnyOrder(record1);
 		}
 
 		@Test
 		public void zeroArgs() {
 			ExitStatus exitStatus = sut.run(List.of(), in, out, err);
-			Assertions.assertThat(exitStatus).isError();
+			assertThat(exitStatus).isError();
 			BDDMockito.then(in).shouldHaveNoMoreInteractions();
 			BDDMockito.then(out).shouldHaveNoMoreInteractions();
 			BDDMockito.then(err).should().send(Records.singleton(Keys.ERROR, Values.ofText("expected 1 parameter: key")));
@@ -1203,11 +1203,11 @@ public class TextModuleTest {
 			Record record1 = Records.builder().entry(Keys.COUNT, Values.ofNumeric(2)).entry(Keys.TEXT, Values.ofText("whatever")).build();
 			BDDMockito.given(in.recv()).willReturn(Optional.of(record1), Optional.empty());
 			ExitStatus exitStatus = sut.run(List.of(), in, out, err);
-			Assertions.assertThat(exitStatus).isSuccess();
+			assertThat(exitStatus).isSuccess();
 			BDDMockito.then(in).shouldHaveNoMoreInteractions();
 			BDDMockito.then(err).shouldHaveNoMoreInteractions();
 			BDDMockito.then(out).should(times(2)).send(records.capture());
-			Assertions.assertThat(records.getAllValues()).containsExactly(
+			assertThat(records.getAllValues()).containsExactly(
 					Records.singleton(Keys.of("header"), Values.withStyle(Values.ofText("count     text      "), Ansi.Style.FG_CYAN)),
 					Records.singleton(Keys.of("row"), Values.withStyle(Values.ofText("2         whatever  "), Ansi.Style.BG_BLUE)));
 		}
@@ -1218,11 +1218,11 @@ public class TextModuleTest {
 			Record record1 = Records.builder().entry(Keys.COUNT, Values.none()).entry(Keys.TEXT, Values.ofText("whatever")).build();
 			BDDMockito.given(in.recv()).willReturn(Optional.of(record1), Optional.empty());
 			ExitStatus exitStatus = sut.run(List.of(), in, out, err);
-			Assertions.assertThat(exitStatus).isSuccess();
+			assertThat(exitStatus).isSuccess();
 			BDDMockito.then(in).shouldHaveNoMoreInteractions();
 			BDDMockito.then(err).shouldHaveNoMoreInteractions();
 			BDDMockito.then(out).should(times(2)).send(records.capture());
-			Assertions.assertThat(records.getAllValues()).containsExactly(
+			assertThat(records.getAllValues()).containsExactly(
 					Records.singleton(Keys.of("header"), Values.withStyle(Values.ofText("count     text      "), Ansi.Style.FG_CYAN)),
 					Records.singleton(Keys.of("row"), Values.withStyle(Values.ofText("          whatever  "), Ansi.Style.BG_BLUE)));
 		}
@@ -1230,7 +1230,7 @@ public class TextModuleTest {
 		@Test
 		public void zeroArgs() {
 			ExitStatus exitStatus = sut.run(List.of("a"), in, out, err);
-			Assertions.assertThat(exitStatus).isError();
+			assertThat(exitStatus).isError();
 			BDDMockito.then(in).shouldHaveNoMoreInteractions();
 			BDDMockito.then(out).shouldHaveNoMoreInteractions();
 			BDDMockito.then(err).should().send(Records.singleton(Keys.ERROR, Values.ofText("expected 0 arguments")));
