+++ pom.xml
@@ -48,6 +48,7 @@
 		<module>modules/terminal</module>
 		<module>modules/fs</module>
 		<module>runtime</module>
+		<module>test-support</module>
 	</modules>
 	<dependencies>
 		<dependency>
