+++ test-support/src/main/java/module-info.java
@@ -21,45 +21,10 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  * SOFTWARE.
  */
-package hosh.testsupport;
+module hosh.test.support {
+	requires hosh.spi;
 
-import org.junit.jupiter.api.extension.AfterTestExecutionCallback;
-import org.junit.jupiter.api.extension.BeforeTestExecutionCallback;
-import org.junit.jupiter.api.extension.Extension;
-import org.junit.jupiter.api.extension.ExtensionContext;
-
-/**
- * To be used to control current thread name or interrupted flag.
- */
-public class WithThread implements Extension, BeforeTestExecutionCallback, AfterTestExecutionCallback {
-
-	private String backup;
-
-	@Override
-	public void afterTestExecution(ExtensionContext context) {
-		Thread.currentThread().setName(backup);
-		Thread.interrupted();
-	}
-
-	@Override
-	public void beforeTestExecution(ExtensionContext context) {
-		backup = Thread.currentThread().getName();
-	}
-
-	public String currentName() {
-		return Thread.currentThread().getName();
-	}
-
-	public void renameTo(String newName) {
-		Thread.currentThread().setName(newName);
-	}
-
-	public boolean isInterrupted() {
-		return Thread.currentThread().isInterrupted();
-	}
-
-	public void interrupt() {
-		Thread.currentThread().interrupt();
-	}
-
+	requires org.assertj.core;
+	requires org.junit.jupiter.api;
+	requires org.mockito;
 }
